+++ commands/moderators/autopush.js
@@ -3,10 +3,10 @@ const prefix   = process.env.PREFIX;
 const settings = require('../../ressources/settings');
 const strings  = require('../../ressources/strings');
 
-const { date }       = require('../../helpers/date')
-const { doPush }     = require('../../functions/doPush');
-const { getResults } = require('../../functions/textures_submission/getResults');
-const { warnUser }   = require('../../helpers/warnUser');
+const { date }            = require('../../helpers/date')
+const { pushTextures }    = require('../../functions/textures/admission/pushTextures')
+const { downloadResults } = require('../../functions/textures/admission/downloadResults')
+const { warnUser }        = require('../../helpers/warnUser')
 
 module.exports = {
 	name: 'autopush',
@@ -16,19 +16,18 @@ module.exports = {
 	syntax: `${prefix}autopush <both/c32/c64>`,
 	example: `${prefix}autopush c32`,
 	async execute(client, message, args) {
-    if(!message.member.hasPermission('ADMINISTRATOR')) return warnUser(message, strings.COMMAND_NO_PERMISSION);
-
-		if (!args.length) return warnUser(message, strings.COMMAND_WRONG_ARGUMENTS_GIVEN);
+    if(!message.member.hasPermission('ADMINISTRATOR')) return warnUser(message, strings.COMMAND_NO_PERMISSION)
+		if (!args.length) return warnUser(message, strings.COMMAND_WRONG_ARGUMENTS_GIVEN)
 
 		if (args[0] == 'both') {
-			await getResults(client, settings.C64_RESULTS);
-			await getResults(client, settings.C32_RESULTS);
+			await downloadResults(client, settings.C64_RESULTS)
+			await downloadResults(client, settings.C32_RESULTS)
 		}
-		else if (args[0] == 'c32') await getResults(client, settings.C32_RESULTS);
-		else if (args[0] == 'c64') await getResults(client, settings.C64_RESULTS);
+		else if (args[0] == 'c32') await downloadResults(client, settings.C32_RESULTS)
+		else if (args[0] == 'c64') await downloadResults(client, settings.C64_RESULTS)
 		else return warnUser(message, strings.COMMAND_WRONG_ARGUMENTS_GIVEN);
 
-		await doPush(`Manual githubPush, executed by: ${message.author.username} (${date()})`);	// Push them trough GitHub
+		await pushTextures(`Manual push, executed by: ${message.author.username} (${date()})`);	// Push them trough GitHub
 
 		return await message.react('✅');
 	}