+++ ressources/settings.js
@@ -14,12 +14,12 @@ const C32_MODERATOR_ID  = '773984348860711003'
 const C32_LOGS          = '798893497204998161' // #logs channel
 const C32_MOD_LOGS      = '776145239358046248' // #mod-log channel
 const C32_COUNCIL_ID    = '775636065338785813' // Council role
-const C32_SUBMIT_1      = '773987409993793546' // #submit-java
-const C32_SUBMIT_1B     = '803652559922266192' // #submit-bedrock
-const C32_SUBMIT_2      = '774220983044669450' // #council-voting
-const C32_SUBMIT_3      = '780507681987100682' // #texture-revote
-const C32_RESULTS       = '780507804317384744' // #texture-results
-const C32_githubPush_FAIL = '779363363158229004' // #council-talk
+
+const C32_SUBMIT_TEXTURES = '849395454390304829' // #submit-textures
+const C32_SUBMIT_COUNCIL  = '774220983044669450' // #council-voting
+const C32_SUBMIT_REVOTE   = '780507681987100682' // #texture-revote
+const C32_RESULTS         = '780507804317384744' // #texture-results
+
 const C32_COUNTER       = '774333964101615637'
 
 // Compliance 64x
@@ -28,12 +28,11 @@ const C64_ID            = '747574286356840609'
 const C64_MODERATOR_ID  = '747595211273797724'
 const C64_LOGS          = '778189351703412767' // #logs channel
 const C64_COUNCIL_ID    = '777642015885754368' // Council role
-const C64_SUBMIT_1      = '747603911677837454' // #submit-java
-const C64_SUBMIT_1B     = '803655327798132757' // #submit-bedrock
-const C64_SUBMIT_2      = '797780897805238282' // #council-voting
-const C64_SUBMIT_3      = '797649484259524648' // #texture-revote
-const C64_RESULTS       = '797648495804678154' // #texture-results
-const C64_githubPush_FAIL = '797780954626261022' // #council-talk
+
+const C64_SUBMIT_TEXTURES = '849398158886567977' // #submit-textures
+const C64_SUBMIT_COUNCIL  = '797780897805238282' // #council-voting
+const C64_SUBMIT_REVOTE   = '797649484259524648' // #texture-revote
+const C64_RESULTS         = '797648495804678154' // #texture-results
 
 // Compliance Addons
 const CADDONS_IMG          = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Discord-Bot/master/ressources/images/Logos/Compliance%20Addons.png'
@@ -71,12 +70,12 @@ module.exports = {
 	C32_MOD_LOGS,
 	C32_MODERATOR_ID,
 	C32_COUNCIL_ID,
-	C32_SUBMIT_1,
-	C32_SUBMIT_1B,
-	C32_SUBMIT_2,
-	C32_SUBMIT_3,
+
+	C32_SUBMIT_TEXTURES,
+	C32_SUBMIT_COUNCIL,
+	C32_SUBMIT_REVOTE,
 	C32_RESULTS,
-	C32_githubPush_FAIL,
+
 	C32_COUNTER,
 
 	C64_MODERATOR_ID,
@@ -84,12 +83,11 @@ module.exports = {
 	C64_ID,
 	C64_LOGS,
 	C64_COUNCIL_ID,
-	C64_SUBMIT_1,
-	C64_SUBMIT_1B,
-	C64_SUBMIT_2,
-	C64_SUBMIT_3,
+
+	C64_SUBMIT_TEXTURES,
+	C64_SUBMIT_COUNCIL,
+	C64_SUBMIT_REVOTE,
 	C64_RESULTS,
-	C64_githubPush_FAIL,
 
 	CADDONS_IMG,
 	CADDONS_ID,